+++ src/engine/GPURenderer/r_glimp.cpp
@@ -157,7 +157,7 @@ static void GLimp_DetectAvailableModes( void )
         return;
     }
     
-    modes = ( SDL_Rect* )SDL_calloc( ( size_t )numSDLModes, sizeof( SDL_Rect ) );
+    modes = ( SDL_Rect* )SDL_calloc( ( uint64 )numSDLModes, sizeof( SDL_Rect ) );
     if( !modes )
     {
         Com_Error( ERR_FATAL, "Out of memory" );
